마진관리 화면 기능 완성입니다. UI좀 확인부탁드려요

+++ app/src/main/java/com/glowstudio/android/blindsjn/feature/foodcost/model/Margin.kt
@@ -23,6 +23,6 @@ data class RecentSale(
     val title: String,
     val price: Int,
     val total_ingredient_price: Int,
-    val margin: Int,
+    val margin: Double,
     val sale_date: String
 )
